+++ crates/pyo3-gated-macros/src/lib.rs
@@ -2,14 +2,14 @@
 
 use proc_macro::TokenStream;
 use quote::quote;
-use syn::{ItemEnum, ItemImpl, ItemStruct, parse_macro_input};
+use syn::{Item, ItemEnum, ItemImpl, ItemStruct, parse_macro_input};
 
 mod args;
 mod attrs;
 mod expand;
 mod paths;
 
-use args::MacroArgs;
+use args::{MacroArgs, ModuleArgs};
 
 #[proc_macro]
 #[doc(hidden)]
@@ -26,6 +26,13 @@ pub fn py_compat_struct(attr: TokenStream, input: TokenStream) -> TokenStream {
     let args = parse_macro_input!(attr as MacroArgs);
     let input_struct = parse_macro_input!(input as ItemStruct);
     expand::expand_struct(args, input_struct).into()
+}
+
+#[proc_macro_attribute]
+pub fn py_compat(attr: TokenStream, input: TokenStream) -> TokenStream {
+    let args = parse_macro_input!(attr as MacroArgs);
+    let item = parse_macro_input!(input as Item);
+    expand::expand_dispatch(args, item).into()
 }
 
 #[proc_macro_attribute]
@@ -47,4 +54,10 @@ pub fn py_compat_fn(attr: TokenStream, input: TokenStream) -> TokenStream {
     let args = parse_macro_input!(attr as MacroArgs);
     let input_fn = parse_macro_input!(input as syn::ItemFn);
     expand::expand_fn(args, input_fn).into()
+}
+
+#[proc_macro]
+pub fn define_py_module(input: TokenStream) -> TokenStream {
+    let args = parse_macro_input!(input as ModuleArgs);
+    expand::expand_module(args).into()
 }